// api/src/types/enums.ts
export enum WalletStatus {
  ACTIVE = 'active',
  INACTIVE = 'inactive',
}

export enum MerchantStatus {
  ACTIVE = 'active',
  INACTIVE = 'inactive',
}

export enum TransactionType {
  CHARGE = 'charge',
  REFUND = 'refund',
}

export enum TransactionStatus {
  COMPLETED = 'completed',
  DECLINED = 'declined',
}

export enum LedgerEntryType {
  CHARGE = 'charge',
  REFUND = 'refund',
}